Database structure change

+++ Projekt/szakdolgozat-app/app/Http/Controllers/AdvertisementController.php
@@ -3,21 +3,31 @@
 namespace App\Http\Controllers;
 
 use App\Models\Advertisement;
+use App\Models\Category;
 use Illuminate\Http\Request;
+use App\Models\County;
+use App\Models\City;
+use Illuminate\Support\Facades\Auth;
+use App\Models\Picture;
 
 class AdvertisementController extends Controller
 {
-    public function __construct()
-    {
-        $this->middleware('CheckRole:admin')->only(['edit', 'update', 'destroy']);
-    }
+    
     /**
      * Display a listing of the resource.
      */
     public function index()
     {
-        $advertisements = Advertisement::all()->sortByDesc('created_at');
-        return view('advertisements.index', compact('advertisements'));
+        $advertisements = Advertisement::with('pictures')->orderByDesc('created_at')->get();
+        return view('advertisements.list', compact('advertisements'));
+    }
+
+    public function ownAdvertisements()
+    {
+        
+        $advertisements = Advertisement::where('user_id', Auth::user()->user_id)->with('pictures')
+        ->orderByDesc('created_at')->get();
+        return view('advertisements.ownList', compact('advertisements'));
     }
 
     /**
@@ -25,8 +35,18 @@ class AdvertisementController extends Controller
      */
     public function create()
     {
-        return view('advertisements.create');
+        $categories = Category::all();
+        $counties = County::all();
+        $cities = [];
+
+        return view('advertisements.create', compact('categories', 'counties', 'cities'));
     }
+    public function getCitiesByCounty($countyId)
+    {
+        $cities = City::where('county_id', $countyId)->get();
+        return response()->json($cities);
+    }
+
 
     /**
      * Store a newly created resource in storage.
@@ -34,34 +54,35 @@ class AdvertisementController extends Controller
     public function store(Request $request)
     {
         $request->validate([
-            'user_id' => 'required',
             'city_id' => 'required',
             'category_id' => 'required',
             'title' => 'required',
             'price' => 'required',
             'description' => 'required',
+            'pictures.*' => 'image|mimes:jpeg,png,jpg,gif,svg',
         ]);
-
+    
         $newAdvertisement = new Advertisement();
-        $newAdvertisement->user_id = $request->user_id;
+        $newAdvertisement->user_id = Auth::user()->user_id;
         $newAdvertisement->city_id = $request->city_id;
         $newAdvertisement->category_id = $request->category_id;
         $newAdvertisement->title = $request->title;
         $newAdvertisement->price = $request->price;
         $newAdvertisement->description = $request->description;
-
-        if($request->has('picture_id'))
-        {
-            $newAdvertisement->picture_id = $request->picture_id;
-        }
-
-        if ($request->has('mobile_number')) 
-        {
+    
+        if ($request->has('mobile_number')) {
             $newAdvertisement->mobile_number = $request->mobile_number;
         }
-
+    
         $newAdvertisement->save();
-
+    
+        if ($request->hasFile('pictures')) {
+            foreach ($request->file('pictures') as $picture) {
+                $path = $picture->store('advertisement_images', 'public');
+                $newAdvertisement->pictures()->create(['src' => $path]);
+            }
+        }
+    
         return redirect()->route('advertisements.index')->with('status', 'Advertisement created successfully!');
     }
 
@@ -85,19 +106,23 @@ class AdvertisementController extends Controller
      */
     public function edit($id)
     {
-        if (auth()->user()->user_id != Advertisement::find($id)->user_id) 
-        {
-            return redirect()->route('advertisements.index')->with('error', 'You can only edit your own advertisements!');
-        }
-
         $advertisement = Advertisement::find($id);
 
-        if (!$advertisement) 
-        {
-            return redirect()->route('advertisements.index')->with('error', 'Advertisement not found!');
+        if (!$advertisement) {
+            return redirect()->route('advertisements.own')->with('error', 'Advertisement not found!');
         }
 
-        return view('advertisements.edit', compact('advertisement'));
+        if (auth()->user()->user_id != $advertisement->user_id) {
+            return redirect()->route('advertisements.own')->with('error', 'You can only edit your own advertisements!');
+        }
+
+        $advertisement->load('pictures');
+
+        $categories = Category::all();
+        $counties = County::all();
+        $cities = City::where('county_id', $advertisement->county_id)->get();
+
+        return view('advertisements.edit', compact('advertisement', 'counties', 'cities', 'categories'));
     }
 
     /**
@@ -107,35 +132,40 @@ class AdvertisementController extends Controller
     {
         if (auth()->user()->user_id != Advertisement::find($id)->user_id) 
         {
-            return redirect()->route('advertisements.index')->with('error', 'You can only edit your own advertisements!');
+            return redirect()->route('advertisements.own')->with('error', 'You can only edit your own advertisements!');
         }
 
         $advertisement = Advertisement::find($id);
+        $this->authorize('update', $advertisement);
 
         if (!$advertisement) 
         {
-            return redirect()->route('advertisements.index')->with('error', 'Advertisement not found!');
+            return redirect()->route('advertisements.own')->with('error', 'Advertisement not found!');
         }
 
         $request->validate([
-            'user_id' => 'required',
             'city_id' => 'required',
             'category_id' => 'required',
             'title' => 'required',
             'price' => 'required',
             'description' => 'required',
+            'pictures.*' => 'image|mimes:jpeg,png,jpg,gif,svg|max:2048',
         ]);
 
-        $advertisement->user_id = $request->user_id;
+        $advertisement->user_id = Auth::user()->user_id;
         $advertisement->city_id = $request->city_id;
         $advertisement->category_id = $request->category_id;
         $advertisement->title = $request->title;
         $advertisement->price = $request->price;
         $advertisement->description = $request->description;
 
-        if($request->has('picture_id'))
+        if ($request->hasFile('pictures')) 
         {
-            $advertisement->picture_id = $request->picture_id;
+            foreach ($request->file('pictures') as $picture) 
+            {
+                $path = $picture->store('advertisement_images', 'public');
+                $advertisement->pictures()->create(['src' => $path]);
+            }
         }
 
         if ($request->has('mobile_number')) 
@@ -145,8 +175,7 @@ class AdvertisementController extends Controller
 
         $advertisement->update();
 
-        return redirect()->route('advertisements.index')->with('status', 'Advertisement updated successfully!');
-
+        return redirect()->route('advertisements.own')->with('status', 'Advertisement updated successfully!');
 
     }
 
@@ -155,20 +184,32 @@ class AdvertisementController extends Controller
      */
     public function destroy($id)
     {
+        
+
         if (auth()->user()->user_id != Advertisement::find($id)->user_id) 
         {
-            return redirect()->route('advertisements.index')->with('error', 'You can only delete your own advertisements!');
+            return redirect()->route('advertisements.own')->with('error', 'You can only delete your own advertisements!');
         }
 
         $advertisement = Advertisement::find($id);
 
+        $this->authorize('delete', $advertisement);
+
         if (!$advertisement) 
         {
-            return redirect()->route('advertisements.index')->with('error', 'Advertisement not found!');
+            return redirect()->route('advertisements.own')->with('error', 'Advertisement not found!');
         }
 
         $advertisement->delete();
 
-        return redirect()->route('advertisements.index')->with('status', 'Advertisement deleted successfully!');
+        return redirect()->route('advertisements.own')->with('status', 'Advertisement deleted successfully!');
+    }
+
+    public function showByCategory($categoryId)
+    {
+        $category = Category::findOrFail($categoryId);
+        $advertisements = $category->advertisements;
+
+        return view('categories.show', compact('advertisements', 'category'));
     }
 }
